# backend/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Cusman CRM API Gateway",
    "description": "Production-grade RESTful API Gateway for Cusman CRM — AI-Powered Customer Growth Platform by DSK Technologies. Supports multi-tenant SaaS integration, WhatsApp messaging, and ERP synchronizations.",
    "version": "1.0.0",
    "contact": {
      "name": "DSK Technologies Support",
      "email": "support@dsktechnologies.com"
    }
  },
  "servers": [
    {
      "url": "https://crm.api.amudhasurabiy.com",
      "description": "Production API Server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Access Token for authorization. Add header 'Authorization: Bearer <TOKEN>'."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "Developer API Key authorization for external automated synchronization (e.g. ERP integration)."
      }
    },
    "schemas": {
      "Customer": {
        "type": "object",
        "required": ["name", "phone"],
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "name": { "type": "string", "example": "Aman Verma" },
          "phone": { "type": "string", "example": "919988776655" },
          "email": { "type": "string", "format": "email", "example": "aman@company.com" },
          "address": { "type": "string", "example": "123 Main St, Central Plaza" },
          "city": { "type": "string", "example": "Coimbatore" },
          "gstNumber": { "type": "string", "example": "33AAAAA1111A1Z1" },
          "company": { "type": "string", "example": "Apex Traders" },
          "tags": { "type": "string", "example": "Retailer,Premium" },
          "birthday": { "type": "string", "format": "date", "example": "1990-05-15" },
          "outstandingAmount": { "type": "number", "format": "float", "example": 1500.50 }
        }
      },
      "Lead": {
        "type": "object",
        "required": ["name", "phone"],
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "name": { "type": "string", "example": "Suresh Kumar" },
          "phone": { "type": "string", "example": "919876543210" },
          "city": { "type": "string", "example": "Bangalore" },
          "company": { "type": "string", "example": "SK Logistics" },
          "tags": { "type": "string", "example": "Lead,Hot" },
          "leadSource": { "type": "string", "example": "SaaS Connector" },
          "leadStage": { "type": "string", "enum": ["New", "Contacted", "Qualified", "Proposal Sent", "Negotiation", "Won", "Lost"], "example": "Qualified" },
          "leadScore": { "type": "string", "enum": ["Hot", "Warm", "Cold"], "example": "Hot" },
          "conversionProbability": { "type": "number", "format": "float", "example": 0.85 }
        }
      },
      "Product": {
        "type": "object",
        "required": ["name", "sku"],
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "name": { "type": "string", "example": "Organic Green Tea" },
          "sku": { "type": "string", "example": "ORG-GRN-TEA" },
          "price": { "type": "number", "format": "float", "example": 199.00 },
          "stock": { "type": "integer", "example": 150 },
          "description": { "type": "string", "example": "Fresh organic green tea leaves directly from farms." },
          "category": { "type": "string", "example": "Beverages" },
          "brand": { "type": "string", "example": "Amudhasurabiy Organics" },
          "imageUrl": { "type": "string", "format": "uri", "example": "https://res.cloudinary.com/demo/image/upload/v1/tea.jpg" },
          "catalogueUrl": { "type": "string", "format": "uri", "example": "https://res.cloudinary.com/demo/raw/upload/v1/tea_catalog.pdf" }
        }
      },
      "Order": {
        "type": "object",
        "required": ["customerName", "phone", "items"],
        "properties": {
          "id": { "type": "integer", "readOnly": true },
          "customerName": { "type": "string", "example": "Aman Verma" },
          "phone": { "type": "string", "example": "919988776655" },
          "city": { "type": "string", "example": "Coimbatore" },
          "totalValue": { "type": "number", "format": "float", "example": 597.00 },
          "status": { "type": "string", "example": "New Order" },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "productId": { "type": "integer", "example": 1 },
                "productName": { "type": "string", "example": "Organic Green Tea" },
                "quantity": { "type": "integer", "example": 3 },
                "unitPrice": { "type": "number", "example": 199.00 }
              }
            }
          }
        }
      }
    }
  },
  "security": [
    { "BearerAuth": [] },
    { "ApiKeyAuth": [] }
  ],
  "paths": {
...
    "/api/customers": {
      "get": {
        "summary": "List all customers (Contacts) for the workspace",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Search by name, phone, email, etc." },
          { "name": "tag", "in": "query", "schema": { "type": "string" }, "description": "Filter by custom tags." }
        ],
        "responses": {
          "200": { "description": "List of customers returned." }
        }
      },
      "post": {
        "summary": "Create a new customer (Contact)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Customer" }
            }
          }
        },
        "responses": {
          "201": { "description": "Customer created successfully." }
        }
      }
    },
...
    "/api/leads": {
      "get": {
        "summary": "List all Active Leads",
        "responses": {
          "200": { "description": "List of active leads." }
        }
      },
      "post": {
        "summary": "Create a new Lead",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Lead" }
            }
          }
        },
        "responses": {
          "201": { "description": "Lead created successfully." }
        }
      }
    },
...
    "/api/products": {
      "get": {
        "summary": "List products and catalogues",
        "responses": {
          "200": { "description": "List of products." }
        }
      },
      "post": {
        "summary": "Create a new Product",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Product" }
            }
          }
        },
        "responses": {
          "201": { "description": "Product created." }
        }
      }
    },
...
    "/api/orders": {
      "get": {
        "summary": "List all sales orders",
        "responses": {
          "200": { "description": "List of sales orders." }
        }
      },
      "post": {
        "summary": "Create a new Sales Order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Order" }
            }
          }
        },
        "responses": {
          "201": { "description": "Sales Order created." }
        }
      }
    },
...
  }
}